Move MessageReader out of request package and up to message package
